requirements.txt
botbuilder-integration-aiohttp>=4.13.0
botbuilder-ai>=4.13.0
botbuilder-core>=4.7.0
aiohttp

